changed returntype int -> long long; added a TESTBED

+++ inc/timers.h
@@ -4,4 +4,4 @@
 long long readTSC(void);	/* not sure if we keep this public */
 void init_timers(int n);
 void stamp_timers(int i);
-int diff_timers(int i, int j);
+long long diff_timers(int i, int j);
